Kill child processes of Sneakers instead of waiting for main process to dissapear
Fixes endless waiting for Sneakers process death in Docker.
For some reason killed Sneakers main process end up being a zombie.
This behaviour impacts only specs run, so does not worth investigating the root of the problem.
Possibly, it can be related to PID 1 problem of Docker. https://blog.phusion.nl/2015/01/20/docker-and-the-pid-1-zombie-reaping-problem/